requirements.txt
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
twilio>=8.0.0
cryptography>=41.0.0
anthropic>=0.25.0
